+++ src/full/Agda/Syntax/Concrete/Fixity.hs
@@ -207,7 +207,7 @@ declaredNames d = case d of
   TypeSig _ x _        -> declaresName x
   FieldSig _ x _       -> declaresName x
   Field _ fs           -> foldMap declaredNames fs
-  FunClause (LHS p [] []) _ _ _
+  FunClause (LHS p [] [] _) _ _ _
     | IdentP (QName x) <- removeSingletonRawAppP p
                        -> declaresName x
   FunClause{}          -> mempty